add leading and actions
