+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'model/post.dart';
+import './demo/listview_demo.dart';
 
 void main() {
   runApp(
@@ -12,6 +12,7 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     // TODO: implement build
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       home: Home(),
       theme: ThemeData(
         primaryColor: Colors.yellow
@@ -21,58 +22,28 @@ class App extends StatelessWidget {
 }
 
 class Home extends StatelessWidget {
-  Widget _listItemBuilder(BuildContext context, int index) {
-    return Container(
-      color: Colors.white,
-      margin: EdgeInsets.all(10.0),
-      child: Column(
-        children: <Widget>[
-          Image.network(posts[index].imageUrl),
-          SizedBox(height: 12.0,),
-          Text(
-            posts[index].title,
-            style: Theme.of(context).textTheme.title,
-          ),
-          Text(
-            posts[index].author,
-            style: Theme.of(context).textTheme.subhead,
-          ),
-          SizedBox(height: 12.0,)
-        ],
-      ),
-    );
-  }
-  @override
-    Widget build(BuildContext context) {
-      // TODO: implement build
-      return Scaffold(
-        backgroundColor: Colors.grey[100],
-        appBar: AppBar(
-          title: Text('Lishaoy.net'),
-          elevation: 6.0,
-        ),
-        body: ListView.builder(
-          itemCount: posts.length,
-          itemBuilder: _listItemBuilder,
-        ),
-      );
-    }
-}
-
-class Hello extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // TODO: implement build
-    return Center(
-      child: Text(
-        'holle',
-        textDirection: TextDirection.ltr,
-        style: TextStyle(
-          fontWeight: FontWeight.bold,
-          fontSize: 40.0,
-          color: Colors.black87,
+    return Scaffold(
+      backgroundColor: Colors.grey[100],
+      appBar: AppBar(
+        leading: IconButton(
+          icon: Icon(Icons.menu),
+          tooltip: 'Navigration',
+          onPressed: () => debugPrint('Navitration Button is pressed'),
         ),
+        title: Text('Lishaoy.net'),
+        actions: <Widget>[
+            IconButton(
+            icon: Icon(Icons.search),
+            tooltip: 'Search',
+            onPressed: () => debugPrint('Search Button is pressed'),
+          )
+        ],
+        elevation: 6.0,
       ),
+      body: null
     );
   }
 }
